Project restructuring and elimination of shared library shenanigans (#33)
* Various project restructuring

- We have moved the GLFW-based implementation of `I_window` into a
  separate library, so that it may better be used by future examples.
- Consolidated pong source.
- Added a basic `CMakePresets.json` for Ninja Multi-Config builds, as I
  tend to use.
- We are no longer pretending in any capacity that this library can be
  consumed as a DLL or .so. The interface(s) were not designed with that
  in mind, and a lot of retrograde work would have to be done to make
  that feasible. `mope_game_engine` is strictly a static library, for
  the forseeable future.
- Consolidated some compile options that we plan to use everywhere, and
  also added warnings.

* Settle warnings caused by addition of `/W4`

+++ src/sprite_renderer.cxx
@@ -95,7 +95,7 @@ void mope::sprite_renderer::render(game_scene& scene, double alpha)
         : detail::component_gatherer<sprite_component, transform_component>::gather(scene))
     {
         sprite.texture.bind();
-        m_shader.set_uniform("u_model", transform.blend(alpha));
+        m_shader.set_uniform("u_model", transform.blend(static_cast<float>(alpha)));
         m_vao.bind();
         ::glDrawElements(GL_TRIANGLE_STRIP, 4, GL_UNSIGNED_BYTE, 0);
     }
